Adapted ruslans changes to allow revision->buildnr caching
It's pretty dumb to rebuild this map every time it is requested, jobs now hold a cache which gets refreshed
only if the user explicitly asks for it.

In addition I added support for multiple builds on a single revision which is a long standing problem with the old
jenkins api. If a user manually rebuilds the same revision we end up with multiple builds with the same rev.
This confuses the deployscript, and deploys will fail due to that.

This change opens up the possibility to deal with this more gracefully