Make final return statement be useful
The final return statement was not useful: it was dead code, and was never
executed, since both the try and except branches of the exception
handling were exiting the method before the final return statement
was reached.

Change the method to follow the structure of the get_data method,
by assigning the return value of urlopen into a variable, which the
final return statement then returns.