Fix for https://github.com/salimfadhley/jenkinsapi/issues/53
Changed retry_function to return 404s directly, as in that
case we have reached the server, so it is not a network or internal
server error, redirect, proxy failure or anything like that.

Artifact.valid will consider the fingerprint valid if the
fingerprint is not available at server (but it can still exist, so
it cannot be claimed as invalid with certainty).