Added system tests.
They also can be called acceptance or functional tests.  I have always
found testing terminology somewhat entangled :)

These test are running against an instance of real jenkins, which is
launched in the beginning of the test suite and terminated in the end.
jenkins.war file is automatically downloaded in case it isn't already
present in jenkinsapi_tests/systests directory.

There are just few tests for now.  This commit is more like a
demonstration of the concept.

Nose (pip install nose) is needed for running the tests.  After
installing it tests can be run with 'nosetests' command from top-level
project directory.