Raise NoBuildData when latestBuild etc. not even set
The Folders plugin uses jobs as a container for folders.
However, a lot of standard job keys, such as lastBuild et al.
are not even present (unlike a never run job where they are
present but set to None)

Since a value of None for latestBuild causes a NoBuildData
exception, the same exception should be fine when not present.